e2e: extract the shared memory scenario spec body
